+++ src/esp_logger/logger_config.h
@@ -22,5 +22,6 @@ struct LoggerConfig {
     UBaseType_t priority = 1;
     LogLevel consoleLogLevel = LogLevel::Debug;
     bool enableSyncTask = true;
+    bool usePrettyJson = true;
     bool usePSRAMBuffers = false;
 };